moved tests from main to tests.rs
